Create domain/managers.py , and moved CustomUserManager from models.py to managers.py Added field to UserProfileSerializer & added read_only for identity_id Make migrations
