+++ UserManagement/accounts/adapters/serializers.py
@@ -5,9 +5,10 @@ from ..domain.models import UserProfile, InstructorRate
 class UserProfileSerializer(serializers.ModelSerializer):
     class Meta:
         model = UserProfile
-        fields = '__all__'  # List all fields or specify a list of fields
+        fields = ['id', 'identity_id', 'email', 'full_name', 'date_of_birth',
+                  'is_admin', 'is_instructor', 'is_active']
         extra_kwargs = {
-            'email': {'read_only': True},  # Make 'email' read-only after creation
+            'identity_id': {'read_only': True},  # Make 'identity_id' read-only after creation
         }
